backend/app/models/image_model.py
import torch
import torch.nn as nn
import numpy as np
import cv2
from PIL import Image

class ImageDeepfakeDetector:
    """
    Image deepfake detection model.
    
    This class will load and run inference with a pre-trained model
    for detecting deepfake images.
    """
    
    def __init__(self, model_path=None):
        """
        Initialize the image deepfake detector.
        
        Args:
            model_path (str, optional): Path to the pre-trained model weights.
        """
        self.model_path = model_path
        self.model = None
        # Model would be loaded here
    
    def preprocess(self, image_data):
        """
        Preprocess image data for model input.
        
        Args:
            image_data (bytes): Raw image data.
            
        Returns:
            torch.Tensor: Preprocessed image data ready for model input.
        """
        # Image preprocessing would go here
        return None
    
    def predict(self, image_data):
        """
        Predict whether image is a deepfake.
        
        Args:
            image_data (bytes): Raw image data.
            
        Returns:
            dict: Prediction results including confidence score.
        """
        # Preprocessing and inference would go here
        
        # Placeholder result
        return {
            "is_deepfake": False,
            "confidence": 0.08,
            "processing_time_ms": 120
        } 