backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.7.0.72
pillow==9.5.0
librosa==0.10.0.post2
transformers==4.29.2
python-dotenv==1.0.0
pytest==7.3.1 